Rename action
Rename to action to `is-dotnet-change-available`.
